hyperonc

Module atom

Structs

RustOpaqueExecError

atom_ref_t

@struct atom_ref_t @ingroup atom_group @brief Refers to an Atom owned by another object @note It is not necessary to free atom_ref_t @note A pointer to atom_t can be passed to any function requiring a pointer to atom_ref_t. @warning atom_ref_t must not be accessed beyond the lifetime of the object which owns the atom it references.

atom_t

@struct atom_t @brief Represents an Atom of any type @ingroup atom_group @note atom_t must be freed with atom_free(), or passed by value to a function that takes ownership of the atom.

atom_vec_t

@struct atom_vec_t @brief Contains a vector (list) of Atoms @ingroup atom_vec_group @note atom_vec_t must be freed with atom_vec_free(), or passed by value to a function that takes ownership of the vec. @warning It is unsafe to directly access the fields of this struct, so accessor functions must be used

bindings_set_t

@brief Represents a set of Bindings frames. Potentially expressing all possible matches produced by a match operarion. @ingroup matching_group

bindings_t

@brief Represents a single Bindings frame, which is a group of mutually-compatible variable <-> atom associations, providing a scope in which variable have definde values @ingroup matching_group

gnd_api_t

@struct gnd_api_t @brief A table of callback functions to implement a Grounded Atom with behavior defined in C @ingroup grounded_atom_group @see atom_gnd @see atom_get_object

gnd_t

@struct gnd_t @brief A struct header that must preface a buffer used as a backing object for a Grounded Atom @ingroup grounded_atom_group @see atom_gnd @see atom_get_object

Enums

atom_type_t

@brief Represents whether an atom is a Symbol, Variable, Expression, or Grounded atom. @ingroup atom_group

exec_error_t

@struct exec_error_t @brief Represents a status used to communicate between Grounded Atom execution and the MeTTa interpreter @ingroup grounded_atom_group @note exec_error_t must be freed with exec_error_free(), or returned from an execute function @note The struct members should never be accessed directly, so accessor functions must be used

Functions

atom_bool

@ingroup atom_group @param[in] b boolean value @return an atom_t for the Bool Grounded atom @note The caller must take ownership responsibility for the returned atom_t

atom_clone

@brief Makes a “deep copy” of an atom. Useful to turn an atom_ref_t into an atom_t @ingroup atom_group @param[in] atom A pointer to an atom_t or an atom_ref_t to clone @return A newly created atom_t for the cloned atom @note The caller must take ownership responsibility for the returned atom_t

atom_eq⚠

@brief Checks if two atom objects represent the same conceptual atom @ingroup atom_group @param[in] a A pointer to an atom_t or an atom_ref_t representing the first atom @param[in] b A pointer to an atom_t or an atom_ref_t representing the second atom @return true is the atoms are conceptually identical

atom_expr⚠

@brief Create a new Expression atom with the specified children atoms @ingroup atom_group @param[in] children A packed buffer of atom_t *, representing the children atoms @param[in] size The number of elements in children @return An atom_t for the Expression atom, containing all ‘children’ atoms @note The caller must take ownership responsibility for the returned atom_t @warning This function takes ownership of all children atoms, so they must not be subsequently accessed

atom_expr_from_vec⚠

@brief Create a new Expression atom with the children contained in an atom_vec_t @ingroup atom_group @param[in] children An atom_vec_t containing all children atoms @return An atom_t for the Expression atom, containing all ‘children’ atoms @note The caller must take ownership responsibility for the returned atom_t @warning This function takes ownership of the children vec, so it must not be subsequently accessed

atom_float

@ingroup atom_group @param[in] f float number @return an atom_t for the Number Grounded atom @note The caller must take ownership responsibility for the returned atom_t

atom_free⚠

@brief Frees an atom and all associated resources @ingroup atom_group @param[in] atom The atom to free

atom_get_children⚠

@brief Provides access to all children atoms within an expression atom @ingroup atom_group @param[in] atom A pointer to an atom_t or an atom_ref_t to access @param[in] callback A function that will be called to return access to each atom @param[in] context A pointer to a caller-defined structure to facilitate communication with the callback function @note This function should only be called with Expression atoms

atom_get_grounded_type

@brief Retrieve the grounded type of a Grounded Atom @ingroup atom_group @param[in] atom A pointer to an atom_t or an atom_ref_t to access @return The atom_t expressing the grounded type of atom @note Grounded Types are themselves atoms used by the MeTTa type system for customized behavior. A grounded type is different from a generic atom type, expressed by atom_type_t. Only Grounded atoms have grounded types. TODO: Where to LINK to learn more?? @note The caller must take ownership responsibility for the returned atom_t

atom_get_metatype⚠

@brief Returns the metatype of an atom @ingroup atom_group @param[in] atom A pointer to an atom_t or an atom_ref_t to inspect @return An atom_type_t indicating the type of atom

atom_get_name

@brief Renders the name of an atom into a text buffer @ingroup atom_group @param[in] atom A pointer to an atom_t or an atom_ref_t to get the name of @param[out] buf A buffer into which the text will be written @param[in] buf_len The maximum allocated size of buf @return The length of the name string, minus the string terminator character. If return_value > buf_len + 1, then the text was not fully written and this function should be called again with a larger buffer.

atom_get_object⚠

@brief Access a pointer to an object backing a Grounded Atom @ingroup atom_group @param[in] atom A pointer to an atom_t or an atom_ref_t to access @return The pointer to the gnd_t orginally used to create the Grounded Atom @note This function is only valid for Grounded Atoms implemented via the HyperonC API @warning The returned pointer must not be accessed after the atom has been freed or modified elsewhere

atom_get_space⚠

@brief Access the space wrapped inside a Grounded atom @ingroup atom_group @see atom_gnd_for_space @param[in] atom A pointer to an atom_t or an atom_ref_t that wraps a Space @return A Space handle to the space inside a grounded atom @note The returned space is borrowed from the atom. It must not be accessed after the atom has been freed or modified elsewhere

atom_gnd

@ingroup atom_group @param[in] gnd A pointer to a buffer to back the new Grounded Atom. See the documentation of gnd_t for more info. @return an atom_t for the Grounded atom @note The caller must take ownership responsibility for the returned atom_t

atom_gnd_for_space

@brief Creates a Grounded Atom referencing a Space @ingroup atom_group @param[in] space A pointer to an space_t for accessing the space @return an atom_t for the Grounded atom @note The caller must take ownership responsibility for the returned atom_t @note This function does not consume the space and the space still must be freed with space_free()

atom_gnd_serialize

@brief Serializes a grounded atom using the given serializer @ingroup serializer_group @param[in] atom A pointer to an atom_t or an atom_ref_t to serialize @param[in] api A table of functions the serialize implementation may call to encode the atom value @param[in] context A caller-defined object to pass to functions in the api, to receive the encoded value(s) @return A serial_result_t indicating whether the serialize operation was successful

atom_int

@ingroup atom_group @param[in] n integer number @return an atom_t for the Number Grounded atom @note The caller must take ownership responsibility for the returned atom_t

atom_is_cgrounded⚠

@brief Check if the atom refers to CGrounded object @ingroup atom_group @param[in] atom A pointer to an atom_t or an atom_ref_t to access @return True if the atom_get_object can be used on this atom without panic

atom_is_null⚠

@brief Returns true if the referenced atom is invalid, otherwise returns false @ingroup atom_group @param[in] atom A pointer to an atom_t or an atom_ref_t to inspect @return true if the referenced atom is invalid, otherwise returns false

atom_iterate⚠

@brief Performs a depth-first exhaustive iteration of an atom and all its children recursively @ingroup atom_group @param[in] atom A pointer to an atom_t or an atom_ref_t to iterate @param[in] callback A function that will be called to return access to each contained atom @param[in] context A pointer to a caller-defined structure to facilitate communication with the callback function @note The first result returned will be the atom itself. Only Expression atoms contain deeper trees to iterate.

atom_match_atom

@brief Matches one atom with another, establishing bindings between them @ingroup matching_group @param[in] a A pointer to an atom_t or an atom_ref_t to match @param[in] b A pointer to another atom_t or an atom_ref_t to match against @return A bindings_set_t representing all variable <-> atom bindings established by the match @note The caller must take ownership responsibility for the returned bindings_set_t, and free it with bindings_set_free()

atom_ref⚠

@brief Create an atom_ref_t that points to another atom you own @ingroup atom_group @param[in] atom The atom to reference @return an atom_ref_t referencing atom @warning The returned atom_ref_t must not be accessed after the atom it refers to has been freed, or after ownership of the original atom has been transferred to another function

atom_ref_null⚠

@brief Create an atom_ref_t that points nothing @ingroup atom_group @return an atom_ref_t referencing nothing

atom_sym⚠

@brief Create a new Symbol atom with the specified name @ingroup atom_group @param[in] name The name of the symbol @return A newly created atom_t for the Symbol Atom @note The caller must take ownership responsibility for the returned atom_t

atom_to_str

@brief Renders a human-readable text description of an atom @ingroup atom_group @param[in] atom A pointer to an atom_t or an atom_ref_t to render @param[out] buf A buffer into which the text will be rendered @param[in] buf_len The maximum allocated size of buf @return The length of the description string, minus the string terminator character. If return_value > buf_len + 1, then the text was not fully rendered and this function should be called again with a larger buffer.

atom_var⚠

@brief Create a new Variable atom with the specified name @ingroup atom_group @param[in] name The name for the newly created Variable atom @return An atom_t for the Variable atom @note The caller must take ownership responsibility for the returned atom_t

atom_var_parse_name⚠

@brief Create a new Variable atom parsing full formatted name of the variable @ingroup atom_group @param[in] name The name for the newly created Variable atom in a format <name>#id @return An atom_t for the Variable atom, returns atom_t::null() when parsing fails @note The caller must take ownership responsibility for the returned atom_t

atom_vec_clone

@brief Creates a new atom_vec_t by cloning an existing atom_vec_t @ingroup atom_vec_group @param[in] vec A pointer to an existing atom_vec_t to clone @return A newly created atom_vec_t @note The caller must take ownership responsibility for the returned atom_vec_t

atom_vec_free⚠

@brief Frees a atom_vec_t @ingroup atom_vec_group @param[in] vec The vec to free

atom_vec_from_array⚠

@brief Creates a new atom_vec_t from a C-style array @ingroup atom_vec_group @param[in] atoms A packed buffer of atom_t *, representing the atoms to put into the vec. @param[in] size The number of elements in atoms @return The newly created atom_vec_t @note The caller must take ownership responsibility for the returned atom_vec_t @warning This function takes ownership of all atoms, so they must not be subsequently accessed or freed

atom_vec_get⚠

@brief Access an atom at a specified index in a vec @ingroup atom_vec_group @param[in] vec The vec in which to access the atom @param[in] idx The index of the element to access @return A reference to the atom at the specified index @note In the event that no atom exists at the specified idx, a null atom_ref_t will be returned @warning The atom referenced by the return value from this function remains owned by the atom_vec_t. It must not be accessed after the vec has been modified or freed.

atom_vec_len⚠

@brief Returns the number of elements in a vec @ingroup atom_vec_group @param[in] vec The vec to be inspected @return The count of the number of elements contained within the vec

atom_vec_new

@brief Creates a new empty atom_vec_t @ingroup atom_vec_group @return The newly created atom_vec_t @note The caller must take ownership responsibility for the returned atom_vec_t

atom_vec_pop⚠

@brief Removes the last element from a vec, and returns it @ingroup atom_vec_group @param[in] vec The vec from which to pop the atom @return The last atom_t contained in the vec @note The caller must take ownership responsibility for the returned atom_t @note In the event that the vec is empty, a null atom_t will be returned

atom_vec_push⚠

@brief Push the atom onto the end of the vec @ingroup atom_vec_group @param[in] vec a pointer to an atom_vec_t to push the atom onto @param[in] atom the atom to push onto the vec @warning This function takes ownership of the supplied atom_t, and it must not be subsequently accessed of freed.

atoms_are_equivalent

@brief Checks if two atoms are alpha equivalent. @ingroup atom_group @param[in] a A pointer to an atom_t or an atom_ref_t representing the first atom @param[in] b A pointer to an atom_t or an atom_ref_t representing the second atom @return true if the atoms can be converted to each other by renaming variables, otherwise false

bindings_add_var_binding

@brief Adds a new variable <-> atom association within a bindings_t @ingroup matching_group @param[in] bindings A pointer to the bindings_t to add the association to @param[in] var The Variable atom (L-value) for the variable <-> atom association @param[in] atom The other atom (R-value) for the variable <-> atom association @return true if the bindings_t has been sucessfully updated, and false if the new association would conflict with existing associations in the Bindings frame @warning This function takes ownership of both the var and atom atoms, so they must not be subsequently accessed or freed

bindings_clone

@brief Makes a “deep copy” of a bindings_t @ingroup matching_group @param[in] bindings A pointer to the bindings_t to clone @return A newly created bindings_t @note The caller must take ownership responsibility for the returned bindings_t

bindings_eq

@brief Checks if two bindings_t objects contain identical associations @ingroup matching_group @param[in] a A pointer to a bindings_t representing the first Bindings frame @param[in] b A pointer to a bindings_t representing the second Bindings frame @return true is the Bindings frames are identical, and false otherwise

bindings_free

@brief Frees a bindings_t @ingroup matching_group @param[in] bindings The bindings_t to free

bindings_is_empty

@brief Checks if a bindings_t contains no associations @ingroup matching_group @param[in] bindings A pointer to the bindings_t to inspect @return true is the Bindings frame contains no associations, and false otherwise

bindings_merge

@brief Merges two bindings_t Bindings frames together into a Bindings Set @ingroup matching_group @param[in] _self The first bindings_t to merge. Ownership of this argument is taken by this function @param[in] other A pointer to the second bindings_t to merge. This argument is not modified @return The bindings_set_t that results from the merge @warning This function takes ownership of the _self argument. After calling this function, the bindings_t passed as _self must not be accessed or freed @note The bindings_set_t returned from this function must be freed with bindings_set_free()

bindings_narrow_vars

@brief Removes all variable associations from a bindings_t except those in the supplied list @ingroup matching_group @param[in] bindings A pointer to the bindings_t to modify @param[in] vars A pointer to a vec of Variable atoms, to specify which variables should remain in the Bindings

bindings_new

@brief Creates a new bindings_t containing no variable <-> atom associations, leaving all variables free to match any atom. @ingroup matching_group @return The new bindings_t @note The caller must take ownership responsibility for the returned bindings_t, and free it with bindings_free()

bindings_resolve

@brief Returns the atom bound to the supplied variable name in the bindings_t @ingroup matching_group @param[in] bindings A pointer to the bindings_t to inspect @param[in] var The Variable atom (L-value) for the variable <-> atom association @return The atom_t representing the atom that corresponds to the specified variable, or a NULL atom_ref_t if the variable is not present. @note The caller must take ownership responsibility for the returned atom_t, if it is not NULL

bindings_set_add_var_binding

@brief Adds a new variable <-> atom association to every Bindings frame in a Bindings set @ingroup matching_group @param[in] set A pointer to the bindings_set_t to modify @param[in] var A pointer to the Variable atom @param[in] value A pointer to the other atom to associate the variable with @note This function is equivalent to calling bindings_add_var_binding for each Bindings frame in the Bindings set, but new Bindings frames may be generated if the new association conflicts with existing associations in existing frames

bindings_set_add_var_equality

@brief Asserts equality between two Variable atoms in a Bindings set @ingroup matching_group @param[in] set A pointer to the bindings_set_t to modify @param[in] a A pointer to the first Variable atom for the variable <-> variable association @param[in] b A pointer to the second Variable atom for the variable <-> variable association @note Both a and b must be Variable atoms

bindings_set_clone

@brief Makes a “deep copy” of a bindings_set_t @ingroup matching_group @param[in] set A pointer to the bindings_set_t to use as the source for the clone @return The new cloned bindings_set_t @note The bindings_set_t returned from this function must be freed with bindings_set_free()

bindings_set_empty

@brief Creates a new bindings_set_t without any Bindings frames. Conceptually this means no valid matches exist @ingroup matching_group @return The empty bindings_set_t @note The bindings_set_t returned from this function must be freed with bindings_set_free()

bindings_set_eq

@brief Checks if two bindings_set_t objects contain identical associations @ingroup matching_group @param[in] a A pointer to a bindings_set_t representing the first Bindings Set @param[in] b A pointer to a bindings_set_t representing the second Bindings Set @return true is the Bindings Sets are identical, and false otherwise

bindings_set_free

@brief Frees a bindings_set_t @ingroup matching_group @param[in] set The bindings_set_t to free

bindings_set_from_bindings

@brief Creates a new bindings_set_t with the specified Bindings frame @ingroup matching_group @param[in] bindings The bindings_t to incorporate into the new Bindings Set. Ownership of this argument is taken by this function @return The new bindings_set_t @warning This function takes ownership of the bindings argument, and it must not be subsequently accessed or freed after calling this function @note The bindings_set_t returned from this function must be freed with bindings_set_free()

bindings_set_is_empty

@brief Checks if a bindings_set_t contains no Bindings frames, and thus indicates no match @ingroup matching_group @param[in] set A pointer to the bindings_set_t to inspect @return true is the Bindings set is empty, and false otherwise

bindings_set_is_single

@brief Checks if a bindings_set_t contains a frame with no associations, and is thus allows variables to take on any value @ingroup matching_group @param[in] set A pointer to the bindings_set_t to inspect @return true is the Bindings set is “single”, and false otherwise @note Perhaps the word “single” is not a good description of its meaning. See https://github.com/trueagi-io/hyperon-experimental/issues/281

bindings_set_iterate

@brief Provides sequential access to each Bindings frame within a Bindings set. @ingroup matching_group @param[in] set A pointer to the bindings_set_t to iterate @param[in] callback A function that will be called to provide access to each Bindings frame @param[in] context A pointer to a caller-defined structure to facilitate communication with the callback function @note The current implementation of this function allows Bindings frames to be modified by the callback function, however this may not be possible in future versions of Hyperon

bindings_set_merge_into

@brief Merges the contents of one Bindings set into another Bindings set @ingroup matching_group @param[in] _self A pointer to the bindings_set_t to modify @param[in] other A pointer to another bindings_set_t @note This function merges each Bindings frame in the _self set with each frame in the other set, using the same logic as bindings_merge. New frames will potentially be created when associations are found to conflict

bindings_set_push

@brief Adds a Bindings frame to an existing bindings_set_t @ingroup matching_group @param[in] set The bindings_set_t to add bindings into @param[in] bindings The bindings_t to incorporate into set. Ownership of this argument is taken by this function @warning This function takes ownership of the bindings argument, and it must not be subsequently accessed or freed after calling this function

bindings_set_single

@brief Creates a new bindings_set_t with one new Bindings frame. Conceptually this means all variables are able to take on any value @ingroup matching_group @return The new bindings_set_t @note The bindings_set_t returned from this function must be freed with bindings_set_free()

bindings_set_to_str

@brief Renders a text description of a bindings_set_t @ingroup matching_group @param[in] set A pointer to the bindings_set_t to render @param[out] buf A buffer into which the text will be rendered @param[in] buf_len The maximum allocated size of buf @return The length of the description string, minus the string terminator character. If return_value > buf_len + 1, then the text was not fully rendered and this function should be called again with a larger buffer.

bindings_to_str

@brief Renders a text description of a bindings_t @ingroup matching_group @param[in] bindings A pointer to the bindings_t to render @param[out] buf A buffer into which the text will be rendered @param[in] buf_len The maximum allocated size of buf @return The length of the description string, minus the string terminator character. If return_value > buf_len + 1, then the text was not fully rendered and this function should be called again with a larger buffer.

bindings_traverse

@brief Iterates each variable <-> Atom association within a bindings_t @ingroup matching_group @param[in] bindings A pointer to the bindings_t to iterate @param[in] callback A function that will be called for each variable <-> atom pair within the Bindings frame @param[in] context A pointer to a caller-defined structure to facilitate communication with the callback function

exec_error_free

@brief Frees an exec_error_t @ingroup grounded_atom_group @param[in] error The exec_error_t to free

exec_error_incorrect_argument

@brief Creates a new exec_error_t representing a “Incorrect Argument” status, telling the MeTTa interpreter that argument was not recognized by the function implementation. @ingroup grounded_atom_group @return The newly created exec_error_t @note The caller must take ownership responsibility for the returned exec_error_t, and ultimately free it with exec_error_free() or return it from an execute function

exec_error_no_err

@brief Creates a new exec_error_t representing a “No Error” status. This is the default interpreter status @ingroup grounded_atom_group @return The newly created exec_error_t @note The caller must take ownership responsibility for the returned exec_error_t, and ultimately free it with exec_error_free() or return it from an execute function

exec_error_no_reduce

@brief Creates a new exec_error_t representing a “Don’t Reduce” status, telling the MeTTa interpreter to process the atoms as they are @ingroup grounded_atom_group @return The newly created exec_error_t @note The caller must take ownership responsibility for the returned exec_error_t, and ultimately free it with exec_error_free() or return it from an execute function

exec_error_runtime

@brief Creates a new exec_error_t representing a runtime error that will halt the MeTTa interpreter @ingroup grounded_atom_group @param[in] message A human-readable error message, for the interpreter to propagate to the user @return The newly created exec_error_t @note The caller must take ownership responsibility for the returned exec_error_t, and ultimately free it with exec_error_free() or return it from an execute function

throw_panic_with_message

@brief Used to throw panic from C Api with message to prevent panics without any info @ingroup grounded_atom_group @param[in] message A human-readable error message which will be used to throw a panic @return ()

Type Aliases

bindings_mut_callback_t

@brief Function signature for a callback providing mutable access to individual Bindings frames within a Bindings Set @ingroup matching_group @param[in] bindings A pointer to the bindings_t. It is ok to call functions that modify the bindings_t within the callback @param[in] context The context state pointer initially passed to the upstream function initiating the callback

c_atom_callback_t

@brief Function signature for a callback providing access to an atom @ingroup atom_group @param[in] atom A reference to the atom. This atom should not be modified or freed by the callback. @param[in] context The context state pointer initially passed to the upstream function initiating the callback.

c_atom_vec_callback_t

@brief Function signature for a callback providing access to an atom_vec_t @ingroup atom_vec_group @param[in] vec The atom_vec_t being provided. This vec should not be modified or freed by the callback. @param[in] context The context state pointer initially passed to the upstream function initiating the callback.

c_var_binding_callback_t

@brief Function signature for a callback providing access to a variable <-> atom pair, associated with a binding @ingroup matching_group @param[in] var A reference to the Variable atom. This atom should not be modified or freed by the callback @param[in] value A reference to the other atom associated with the variable. This atom should not be modified or freed by the callback @param[in] context The context state pointer initially passed to the upstream function initiating the callback