hyperonc::atom

Function atom_gnd_serialize

#[unsafe(no_mangle)]
pub extern "C" fn atom_gnd_serialize(
    atom: *const atom_ref_t,
    api: *const serializer_api_t,
    context: *mut c_void,
) -> serial_result_t

@brief Serializes a grounded atom using the given serializer @ingroup serializer_group @param[in] atom A pointer to an atom_t or an atom_ref_t to serialize @param[in] api A table of functions the serialize implementation may call to encode the atom value @param[in] context A caller-defined object to pass to functions in the api, to receive the encoded value(s) @return A serial_result_t indicating whether the serialize operation was successful