hyperonc::atom

Function bindings_narrow_vars

#[no_mangle]
pub extern "C" fn bindings_narrow_vars(
    bindings: *mut bindings_t,
    vars: *const atom_vec_t,
)

@brief Removes all variable associations from a bindings_t except those in the supplied list @ingroup matching_group @param[in] bindings A pointer to the bindings_t to modify @param[in] vars A pointer to a vec of Variable atoms, to specify which variables should remain in the Bindings