hyperonc::atom

Type Alias bindings_mut_callback_t

pub type bindings_mut_callback_t = extern "C" fn(bindings: *mut bindings_t, context: *mut c_void);

@brief Function signature for a callback providing mutable access to individual Bindings frames within a Bindings Set @ingroup matching_group @param[in] bindings A pointer to the bindings_t. It is ok to call functions that modify the bindings_t within the callback @param[in] context The context state pointer initially passed to the upstream function initiating the callback