hyperonc

Module metta

Structs

env_builder_t

@brief Represents an environment initialization, in progress @ingroup environment_group @note env_builder_t must be given to environment_init_finish() to properly release it

metta_t

@brief A top-level MeTTa runner @ingroup interpreter_group @note A metta_t must be freed with metta_free() @see metta_new @see metta_free

run_context_t

@brief An interface object providing access to the MeTTa run interface @ingroup interpreter_group

runner_state_t

@brief Represents the state of an in-flight MeTTa execution run @ingroup interpreter_group @note A runner_state_t is initially created by runner_state_new_with_parser(). Each call to metta_run_step(), in a loop, advances the evaluation progress by some amount. When the runner operations have fully resolved, runner_state_is_complete() will return true. Ownership of the runner_state_t must ultimately be released with runner_state_free().

sexpr_parser_t

@brief Represents an S-Expression Parser state machine, to parse input text into an Atom @ingroup tokenizer_and_parser_group @note sexpr_parser_t objects must be freed with sexpr_parser_free()

step_result_t

@brief Contains the state for an in-flight interpreter operation @ingroup interpreter_group @note The step_result_t API is very low-level; It provides direct access to the interpreter without a an environment, parser, or tokenizer. Usually the runner_state_t API is a better choice for executing MeTTa code in most situations. @note A step_result_t is initially created by interpret_init(). Each call to interpret_step(), in a loop, consumes the step_result_t and creates a new one. When the interpreter operation has fully resolved, step_get_result() can provide the final results. Ownership of the step_result_t must ultimately be released with step_get_result(). @see interpret_init @see interpret_step @see step_get_result

syntax_node_t

@brief Represents a component in a syntax tree created by parsing MeTTa code @ingroup tokenizer_and_parser_group @note syntax_node_t objects must be freed with syntax_node_free()

token_api_t

@struct token_api_t @brief A table of callback functions to implement custom atom parsing @ingroup tokenizer_and_parser_group @see tokenizer_register_token

tokenizer_t

@brief Represents a handle to a Tokenizer, capable of recognizing meaningful Token substrings in text @ingroup tokenizer_and_parser_group @note tokenizer_t handles must be freed with tokenizer_free()

Enums

syntax_node_type_t

@brief The type of language construct respresented by a syntax_node_t @ingroup tokenizer_and_parser_group

Functions

ATOM_TYPE_ATOM

@brief Creates a Symbol atom for the special MeTTa symbol: “Atom”, used to indicate that an atom’s type is a generic atom @ingroup metta_language_group @return The atom_t representing the Symbol atom @note The returned atom_t must be freed with atom_free()

ATOM_TYPE_BOOL

@brief Creates an atom used to indicate that an atom’s type is a Bool type. @ingroup metta_language_group @return The atom_t representing the atom @note The returned atom_t must be freed with atom_free()

ATOM_TYPE_EXPRESSION

@brief Creates a Symbol atom for the special MeTTa symbol: “Expression”, used to indicate that an atom’s type is an expression atom @ingroup metta_language_group @return The atom_t representing the Symbol atom @note The returned atom_t must be freed with atom_free()

ATOM_TYPE_GROUNDED

@brief Creates a Symbol atom for the special MeTTa symbol: “Grounded”, used to indicate that an atom’s type is a grounded atom @ingroup metta_language_group @return The atom_t representing the Symbol atom @note The returned atom_t must be freed with atom_free()

ATOM_TYPE_GROUNDED_SPACE

@brief Creates a Symbol atom for the special MeTTa symbol used to indicate that an atom’s type is a wrapper around a Space @ingroup metta_language_group @return The atom_t representing the Symbol atom @note The returned atom_t must be freed with atom_free()

ATOM_TYPE_NUMBER

@brief Creates an atom used to indicate that an atom’s type is a Number type. @ingroup metta_language_group @return The atom_t representing the atom @note The returned atom_t must be freed with atom_free()

ATOM_TYPE_STRING

@brief Creates an atom used to indicate that an atom’s type is a String type. @ingroup metta_language_group @return The atom_t representing the atom @note The returned atom_t must be freed with atom_free()

ATOM_TYPE_SYMBOL

@brief Creates a Symbol atom for the special MeTTa symbol: “Symbol”, used to indicate that an atom’s type is a symbol atom @ingroup metta_language_group @return The atom_t representing the Symbol atom @note The returned atom_t must be freed with atom_free()

ATOM_TYPE_TYPE

@brief Creates a Symbol atom for the special MeTTa symbol: “Type”, used to indicate that an atom represents the type of another atom @ingroup metta_language_group @return The atom_t representing the Symbol atom @note The returned atom_t must be freed with atom_free()

ATOM_TYPE_UNDEFINED

@brief Creates a Symbol atom for the special MeTTa symbol: “%Undefined%” @ingroup metta_language_group @return The atom_t representing the Symbol atom @note The returned atom_t must be freed with atom_free()

ATOM_TYPE_UNIT

@brief Creates an atom used to indicate that an atom’s type is a unit type. @ingroup metta_language_group @return The atom_t representing the atom @note The returned atom_t must be freed with atom_free()

ATOM_TYPE_VARIABLE

@brief Creates a Symbol atom for the special MeTTa symbol: “Variable”, used to indicate that an atom’s type is a variable atom @ingroup metta_language_group @return The atom_t representing the Symbol atom @note The returned atom_t must be freed with atom_free()

EMPTY_ATOM

@brief Creates a Symbol atom for the special MeTTa symbol used to indicate empty results returned by function. @ingroup metta_language_group @return The atom_t representing the Empty atom @note The returned atom_t must be freed with atom_free()

METTA_ATOM

@brief Creates a Symbol atom for the special MeTTa symbol used to indicate calling MeTTa interpreter. @ingroup metta_language_group @return The atom_t representing the interpret atom @note The returned atom_t must be freed with atom_free()

UNIT_ATOM

@brief Creates an atom used to return from functions which are not supposed to return results (print for example). @ingroup metta_language_group @return The atom_t representing the Unit atom @note The returned atom_t must be freed with atom_free()

atom_error_message

@brief Renders the text message from an error expression atom into a buffer @ingroup metta_language_group @param[in] atom The error expression atom from which to extract the error message @param[out] buf A buffer into which the text will be rendered @param[in] buf_len The maximum allocated size of buf @return The length of the message string, minus the string terminator character. If return_value > buf_len + 1, then the text was not fully rendered and this function should be called again with a larger buffer. @warning The atom argument must be an error expression, otherwise this function will panic

atom_is_error

@brief Checks if an atom is a MeTTa error expression @ingroup metta_language_group @param[in] atom A pointer to an atom_t or an atom_ref_t representing the atom to check @return true is the atom is a MeTTa error expression

check_type

@brief Checks whether Atom atom has Type typ in context of space @ingroup metta_language_group @param[in] space A pointer to the space_t representing the space context in which to perform the check @param[in] atom A pointer to the atom_t or atom_ref_t representing the atom whose Type the function will check @param[in] typ A pointer to the atom_t or atom_ref_t representing the type to check against @return true if the Atom’s Type is a match, otherwise false @note This function can be used for a simple type check when there is no need to know type parameters

env_builder_create_config_dir

@brief Sets whether the config dir should be created if it doesn’t already exist @ingroup environment_group @param[in] builder A pointer to the in-process environment builder state @param[in] should_create Whether the directory will be created. Defaults to true

env_builder_init_common_env

@brief Finishes initialization of the common environment @ingroup environment_group @param[in] builder The in-process environment builder state to install as the common environment @return True if the environment was sucessfully initialized with the provided builder state. False if the environment had already been initialized by a prior call

env_builder_push_fs_module_format

@brief Adds logic to interpret a foreign format for MeTTa modules loaded from the file system @ingroup environment_group @param[in] builder A pointer to the in-process environment builder state @param[in] format A pointer to a user-defined structure to store information related to this format

env_builder_push_include_path

@brief Adds a directory to search for module imports @ingroup environment_group @param[in] builder A pointer to the in-process environment builder state @param[in] path A C-style string specifying a path to a directory, to search for modules to load @note The paths will be searched in the order they are added to the `env_builder_t``

env_builder_set_config_dir

@brief Sets the config directory for the environment. @ingroup environment_group @param[in] builder A pointer to the in-process environment builder state @param[in] path A C-style string specifying a path to the config directory

env_builder_set_default_config_dir

@brief Sets the default config directory for the environment. @ingroup environment_group @param[in] builder A pointer to the in-process environment builder state

env_builder_set_is_test

@brief Configures the environment for use in unit testing @ingroup environment_group @param[in] builder A pointer to the in-process environment builder state @param[in] is_test True if the environment is a unit-test environment, False otherwise

env_builder_set_working_dir

@brief Sets the working directory for the environment @ingroup environment_group @param[in] builder A pointer to the in-process environment builder state @param[in] path A C-string specifying a path to a working directory, to search for modules to load. Passing NULL will unset the working directory @note This working directory is not required to be the same as the process working directory, and it will not change as the process’ working directory is changed

env_builder_start

@brief Begins initialization of an environment @ingroup environment_group @return The env_builder_t object representing the in-process environment initialization @note The env_builder_t must be passed to either env_builder_init_common_env or metta_new_with_space in order to properly deallocate it

env_builder_use_default

@brief Creates an env_builder_t to specify that the default common environment should be used @ingroup environment_group @return The env_builder_t object specifying the common environment @note This function exists to supply an argument to metta_new_with_space when no special behavior is desired @note The env_builder_t must be passed to metta_new_with_space

env_builder_use_test_env

@brief A convenience to create an env_builder_t, to specify that a unit-test environment should be used @ingroup environment_group @return The env_builder_t object specifying the unit test environment @note This function exists to supply an argument to metta_new_with_space when performing unit testing @note The env_builder_t must be passed to metta_new_with_space

environment_config_dir

@brief Renders the config_dir path from the common environment into a text buffer @ingroup environment_group @param[out] buf A buffer into which the text will be written @param[in] buf_len The maximum allocated size of buf @return The length of the path string, minus the string terminator character. If return_value > buf_len + 1, then the text was not fully written and this function should be called again with a larger buffer. This function will return 0 if there is no config_dir.

get_atom_types

@brief Provides all types for atom in the context of space @ingroup metta_language_group @param[in] space A pointer to the space_t representing the space context in which to access the Atom’s types @param[in] atom A pointer to the atom_t or atom_ref_t representing the atom whose Types the function will access @param[in] callback A function that will be called to provide a vector of all type atoms associated with the atom argument atom @param[in] context A pointer to a caller-defined structure to facilitate communication with the callback function @note %Undefined% will be provided if atom has no type assigned. An empty vector will be provided if atom is a function call but expected types of arguments are not compatible with passed values

interpret_init

@brief Initializes an interpreter operation and take the initial step @ingroup interpreter_group @param[in] space A pointer to the Space in which to perform the operation @param[in] expr A pointer to an atom_t or atom_ref_t Expression atom to interpret @return A step_result_t representing the outcome from the initial step @note The returned value may represent an error, an immediate result value, or it may be necessary to call interpret_step() in a loop to fully evaluate the execution plan. Ultimately step_get_result() must be called to release the returned step_result_t

interpret_step

@brief Takes a subsequent step in an in-flight interpreter operation @ingroup interpreter_group @param[in] step The existing state for the in-flight interpreter operation @return A new step_result_t representing the outcome from the step

metta_clone_handle

@brief Clones a metta_t handle @ingroup interpreter_group @param[in] metta The handle to clone @return The newly cloned metta_t handle, pointing to the same underlying runner @note The caller must take ownership responsibility for the returned metta_t, and free it with metta_free()

metta_eq

@brief Compares two metta_t handles to test whether the referenced MeTTa runner is the same @ingroup interpreter_group @param[in] a A pointer to the first runner handle @param[in] b A pointer to the first runner handle @return True if the two handles reference the same runner, otherwise False

metta_err_str

@brief Returns the error string associated with the last metta_run, metta_evaluate_atom, or metta_load_module call @ingroup interpreter_group @param[in] metta A pointer to the MeTTa handle @return A pointer to the C-string containing the error that occurred, or NULL if no error occurred @warning The returned pointer should NOT be freed. It must never be accessed after the metta_t has been freed, or any subsequent call to metta_run, metta_evaluate_atom, or metta_load_module has been made.

metta_evaluate_atom

@brief Runs the MeTTa runner to evaluate an input Atom @ingroup interpreter_group @param[in] metta A pointer to the runner handle @param[in] atom The atom_t representing the atom to evaluate @param[in] callback A function that will be called to provide a vector of atoms produced by the evaluation @param[in] context A pointer to a caller-defined structure to facilitate communication with the callback function @note If this function encounters an error, the callback will not be called and the error may be accessed with metta_err_str() @warning This function takes ownership of the provided atom_t, so it must not be subsequently accessed or freed

metta_free

@brief Frees a metta_t handle @ingroup interpreter_group @param[in] metta The handle to free @note The underlying runner may be deallocated if all handles that refer to it have been freed, otherwise the runner itself won’t be freed

metta_get_module_space

@brief Returns the Space for a loaded module @ingroup interpreter_group @param[in] metta A pointer to the handle specifying the runner into which to load the module @param[in] mod_id The module_id_t of the loaded module to access @return The space_t for the module’s Space @note The returned space_t must be freed with space_free

metta_load_module_at_path

@brief Loads a module into the runner from a module resource at a file system path @ingroup interpreter_group @param[in] metta A pointer to the handle specifying the runner into which to load the module @param[in] path A C-string specifying the path from which to load the module @param[in] name A C-string specifying a name for the module, or NULL if you want the module to be private and unable to be loaded by name @return The module_id_t for the loaded module, or invalid if there was an error @note This function effectively bypasses the catalog, for situations where you wish to load a specific module from disk @note If this function encounters an error, the error may be accessed with metta_err_str()

metta_load_module_direct

@brief Loads a module directly into the runner, from a module_loader_t @ingroup interpreter_group @param[in] metta_ref A pointer to the handle specifying the runner into which to load the module @param[in] name_ref A C-string specifying a name for the module @param[in] loader_mov The module_loader_t instance to load the module @return The module_id_t for the loaded module, or invalid if there was an error @note This function might be useful to provide MeTTa modules that are built-in as part of your application @note If this function encounters an error, the error may be accessed with metta_err_str()

metta_new

@brief Creates a new top-level MeTTa Runner, with only the Rust stdlib loaded @ingroup interpreter_group @return A metta_t handle to the newly created runner @note The caller must take ownership responsibility for the returned metta_t, and free it with metta_free()

metta_new_core

@brief Creates a new core MeTTa runner, with no loaded stdlib nor initialization @ingroup interpreter_group @param[in] space A pointer to a handle for the Space for use as the space of the top-level module @param[in] environment An env_builder_t handle to configure the environment to use @return A metta_t handle to the newly created runner @note The caller must take ownership responsibility for the returned metta_t, and free it with metta_free() @note This function does not load any stdlib, nor does it run the init.metta file from the environment

metta_new_with_stdlib_loader

@brief Creates a new top-level MeTTa Runner, with the specified stdlib module loaded @ingroup interpreter_group @param[in] space_ref A pointer to a handle for the Space for use in the Runner’s top-level module @param[in] env_builder_mov An env_builder_t handle to configure the environment to use @param[in] stdlib_loader_mov Stdlib loader implemented in C code. Pass NULL to use the default stdlib @return A metta_t handle to the newly created Runner @note The caller must take ownership responsibility for the returned metta_t, and free it with metta_free() @note Most callers can simply call metta_new. This function is provided to support languages with their own stdlib, that needs to be loaded before the init.metta file is run

metta_run

@brief Runs the MeTTa runner until the input text has been fully parsed and evaluated @ingroup interpreter_group @param[in] metta A pointer to the runner handle @param[in] parser An S-Expression Parser containing the MeTTa text @param[in] callback A function that will be called to provide a vector of atoms produced by the evaluation @param[in] context A pointer to a caller-defined structure to facilitate communication with the callback function @note If this function encounters an error, the callback will not be called and the error may be accessed with metta_err_str() @warning Ownership of the provided parser will be taken by this function, so it must not be subsequently accessed nor freed.

metta_space

@brief Provides access to the Space of the runner’s top-level module @ingroup interpreter_group @param[in] metta A pointer to the runner handle @return A Space handle, to access the Space of the runner’s top-level module @note The caller must take ownership responsibility for the returned space_t and free it with space_free()

metta_tokenizer

@brief Provides access to the Tokenizer of the runner’s top-level module @ingroup interpreter_group @param[in] metta A pointer to the runner handle @return A Tokenizer handle, to access the Tokenizer of the runner’s top-level module @note The caller must take ownership responsibility for the returned tokenizer_t and free it with tokenizer_free()

metta_working_dir

@brief Renders the working directory of the runner’s environment into a buffer @ingroup interpreter_group @param[in] metta A pointer to the runner handle @param[out] buf A buffer into which the path will be rendered @param[in] buf_len The maximum allocated size of buf @return The length of the path string, minus the string terminator character. Returns 0 if the runner’s environment has no working directory. @note If return_value > buf_len + 1, then the text was not fully rendered and this function should be called again with a larger buffer.

run_context_get_metta

@brief Returns a pointer to the metta_t runner that a run context is executing within @ingroup interpreter_group @param[in] run_context A pointer to the run_context_t to access the runner API @return A metta_t handle to access the runner @note The returned metta_t handle is must be freed using metta_free

run_context_get_space

@brief Provides access to the Space of the currently running module @ingroup interpreter_group @param[in] run_context A pointer to the run_context_t to access the runner API @return A Space handle, to access the Space of the currently running module @note The caller must take ownership responsibility for the returned space_t and free it with space_free()

run_context_get_tokenizer

@brief Provides access to the Tokenizer of the currently running module @ingroup interpreter_group @param[in] run_context A pointer to the run_context_t to access the runner API @return A Tokenizer handle, to access the Tokenizer of the currently running module @note The caller must take ownership responsibility for the returned tokenizer_t and free it with tokenizer_free()

run_context_import_dependency

@brief Imports a dependency module into the currently running module. This is “import *” behavior @ingroup module_group @param[in] run_context A pointer to the run_context_t to access the runner API @param[in] mod_id The loaded module_id_t of the module to import

run_context_init_self_module

@brief Called within a module loader function to initialize the new module @ingroup module_group @param[in] run_context A pointer to the run_context_t to access the runner API @param[in] space A pointer to a handle for the new module’s space @param[in] resource_dir_path A C-style string specifying a file system path to use as the module’s resource directory. Passing NULL means the module does not have a resource directory @note this function must be called exactly once within a module loader function

run_context_load_module

@brief Resolves a module name in the context of a running module, and loads that module if it’s not already loaded @ingroup module_group @param[in] run_context A pointer to the run_context_t to access the runner API @param[in] name A C-style string containing the module name @return The module_id_t for the loaded module, or invalid if there was an error

run_context_push_parser

@brief Appends the parser to the Run Context’s queue of input to run @ingroup interpreter_group @param[in] run_context A pointer to the run_context_t to access the runner API @param[in] parser An S-Expression Parser containing the MeTTa source code to execute

run_context_raise_error

@brief Sets a runtime error @ingroup interpreter_group @param[in] run_context A pointer to the run_context_t to access the runner API @param[in] message A C-string specifying an error message @note Raising an error through this function will cause the MeTTa interpreter to take an error pathway

runner_state_current_results

@brief Accesses the current in-flight results in the runner_state_t @ingroup interpreter_group @param[in] state The runner_state_t within which to preview results @param[in] callback A function that will be called to provide a vector of atoms produced by the evaluation @param[in] context A pointer to a caller-defined structure to facilitate communication with the callback function @warning The provided results will be overwritten by the next call to metta_run_step, so the caller must clone the result atoms if they are needed for an extended period of time

runner_state_err_str

@brief Returns the error string associated with the last runner_state_step @ingroup interpreter_group @param[in] state A pointer to the runner state @return A pointer to the C-string containing the error that occurred, or NULL if no error occurred @warning The returned pointer should NOT be freed. It must never be accessed after the runner_state_t has been freed, or any subsequent call to runner_state_step has been made.

runner_state_free

@brief Frees a runner_state_t @ingroup interpreter_group @param[in] node The runner_state_t to free

runner_state_is_complete

@brief Returns whether or not the runner_state_t has completed all outstanding work @ingroup interpreter_group @param[in] state The runner_state_t to inspect @return true if the runner has already concluded, or false if there is more work to do

runner_state_new_with_atoms

@brief Creates a runner_state_t, to use for step-wise execution of a list of atoms @ingroup interpreter_group @param[in] metta A pointer to the runner handle in which to perform the run @param[in] atoms A pointer to an atom_vec_t containing the atoms to run @return The newly created runner_state_t, which can begin evaluating MeTTa code @warning The referenced atoms atom_vec_t must not be modified nor freed while the runner_state_t remains active @note The returned runner_state_t handle must be freed with runner_state_free()

runner_state_new_with_parser

@brief Creates a runner_state_t, to use for step-wise execution of MeTTa text @ingroup interpreter_group @param[in] metta A pointer to the runner handle in which to perform the run @param[in] parser An S-Expression Parser containing the MeTTa text @return The newly created runner_state_t, which can begin evaluating MeTTa code @warning Ownership of the provided parser will be taken by this function, so it must not be subsequently accessed nor freed. @note The returned runner_state_t handle must be freed with runner_state_free()

runner_state_step

@brief Runs one step of the runner @ingroup interpreter_group @param[in] state A pointer to the in-flight runner state @note If this function encounters an error, the error may be accessed with runner_state_err_str()

runner_state_to_str

@brief Renders a text description of a runner_state_t into a buffer @ingroup interpreter_group @param[in] state A pointer to a runner_state_t to render @param[out] buf A buffer into which the text will be rendered @param[in] buf_len The maximum allocated size of buf @return The length of the description string, minus the string terminator character. If return_value > buf_len + 1, then the text was not fully rendered and this function should be called again with a larger buffer.

sexpr_parser_err_str

@brief Returns the error string associated with the last sexpr_parser_parse call @ingroup tokenizer_and_parser_group @param[in] parser A pointer to the Parser, which is associated with the text to parse @return A pointer to the C-string containing the parse error that occurred, or NULL if no parse error occurred @warning The returned pointer should NOT be freed. It must never be accessed after the sexpr_parser_t has been freed, or any subsequent call to sexpr_parser_parse or sexpr_parser_parse_to_syntax_tree has been made.

sexpr_parser_free

@brief Frees an S-Expression Parser @ingroup tokenizer_and_parser_group @param[in] parser The sexpr_parser_t handle to free

sexpr_parser_new

@brief Creates a new S-Expression Parser @ingroup tokenizer_and_parser_group @param[in] text A C-style string containing the input text to parse @return The new sexpr_parser_t, ready to parse the text @note The returned sexpr_parser_t must be freed with sexpr_parser_free() or passed to another function that takes ownership @warning The returned sexpr_parser_t borrows a reference to the text, so the returned sexpr_parser_t must be freed before the text is freed or allowed to go out of scope.

sexpr_parser_parse

@brief Parses the text associated with an sexpr_parser_t, and creates the corresponding Atom @ingroup tokenizer_and_parser_group @param[in] parser A pointer to the Parser, which is associated with the text to parse @param[in] tokenizer A pointer to the Tokenizer, to use to interpret atoms within the expression @return The new atom_t, which may be an Expression atom with many child atoms. Returns a none atom if parsing is finished, or an error expression atom if a parse error occurred. @note The caller must take ownership responsibility for the returned atom_t, and ultimately free it with atom_free() or pass it to another function that takes ownership responsibility @note If this function encounters an error, the error may be accessed with sexpr_parser_err_str()

sexpr_parser_parse_to_syntax_tree

@brief Parses the text associated with an sexpr_parser_t, and creates a syntax tree @ingroup tokenizer_and_parser_group @param[in] parser A pointer to the Parser, which is associated with the text to parse @return The new syntax_node_t representing the root of the parsed tree @note The caller must take ownership responsibility for the returned syntax_node_t, and ultimately free it with syntax_node_free()

step_get_result

@brief Consumes a step_result_t and provides the ultimate outcome of a MeTTa interpreter session @ingroup interpreter_group @param[in] step A pointer to a step_result_t to render @param[in] callback A function that will be called to provide a vector of all atoms resulting from the interpreter session @param[in] context A pointer to a caller-defined structure to facilitate communication with the callback function

step_has_next

@brief Examines a step_result_t to determine if more work is needed @ingroup interpreter_group @param[in] step A pointer to the step_result_t representing the in-flight interpreter operation @return true if the operation plan indicates that more work is needed to finalize results, otherwise false

step_to_str

@brief Renders a text description of a step_result_t into a buffer @ingroup interpreter_group @param[in] step A pointer to a step_result_t to render @param[out] buf A buffer into which the text will be rendered @param[in] buf_len The maximum allocated size of buf @return The length of the description string, minus the string terminator character. If return_value > buf_len + 1, then the text was not fully rendered and this function should be called again with a larger buffer.

syntax_node_clone

@brief Creates a deep copy of a syntax_node_t @ingroup tokenizer_and_parser_group @param[in] node A pointer to the syntax_node_t @return The syntax_node_t representing the cloned syntax node @note The caller must take ownership responsibility for the returned syntax_node_t, and ultimately free it with syntax_node_free()

syntax_node_free

@brief Frees a syntax_node_t @ingroup tokenizer_and_parser_group @param[in] node The syntax_node_t to free

syntax_node_is_leaf

@brief Returns true if a syntax node is a leaf (has no children) and false otherwise @ingroup tokenizer_and_parser_group @param[in] node A pointer to the syntax_node_t @return The boolean value indicating if the node is a leaf

syntax_node_is_null

@brief Returns true if a syntax node represents the end of the stream @ingroup tokenizer_and_parser_group @param[in] node A pointer to the syntax_node_t @return The boolean value indicating if the node is a a null node

syntax_node_iterate

@brief Performs a depth-first iteration of all child syntax nodes within a syntax tree @ingroup tokenizer_and_parser_group @param[in] node A pointer to the top-level syntax_node_t representing the syntax tree @param[in] callback A function that will be called to provide a vector of all type atoms associated with the atom argument atom @param[in] context A pointer to a caller-defined structure to facilitate communication with the callback function

syntax_node_src_range

@brief Returns the beginning and end positions in the parsed source of the text represented by the syntax node @ingroup tokenizer_and_parser_group @param[in] node A pointer to the syntax_node_t @param[out] range_start A pointer to a value, into which the starting offset of the range will be written @param[out] range_end A pointer to a value, into which the ending offset of the range will be written

syntax_node_type

@brief Returns the type of a syntax_node_t @ingroup tokenizer_and_parser_group @param[in] node A pointer to the syntax_node_t @return The syntax_node_type_t representing the type of the syntax node

tokenizer_clone

@brief Performs a “deep copy” of a Tokenizer @ingroup tokenizer_and_parser_group @param[in] tokenizer A pointer to the Tokenizer to clone @return The new Tokenizer, containing all registered Tokens belonging to the original Tokenizer @note The returned tokenizer_t must be freed with tokenizer_free()

tokenizer_free

@brief Frees a Tokenizer handle @ingroup tokenizer_and_parser_group @param[in] tokenizer The tokenizer_t handle to free @note When the last tokenizer_t handle for an underlying Tokenizer has been freed, then the Tokenizer will be deallocated

tokenizer_new

@brief Creates a new Tokenizer, without any registered Tokens @ingroup tokenizer_and_parser_group @return an tokenizer_t handle to access the newly created Tokenizer @note The returned tokenizer_t handle must be freed with tokenizer_free()

tokenizer_register_token

@brief Registers a new custom Token in a Tokenizer @ingroup tokenizer_and_parser_group @param[in] tokenizer A pointer to the Tokenizer in which to register the Token @param[in] regex A regular expression to match the incoming text, triggering this token to generate a new atom @param[in] api A table of functions to manage the token @param[in] context A caller-defined structure to communicate any state necessary to implement the Token parser @note Hyperon uses the Rust RegEx engine and syntax, documented here.

validate_atom

@brief Checks whether atom is correctly typed @ingroup metta_language_group @param[in] space A pointer to the space_t representing the space context in which to perform the check @param[in] atom A pointer to the atom_t or atom_ref_t representing the atom whose Type the function will check @return true if the Atom is correctly typed, otherwise false @note This function can be used to check if function arguments have correct types

Type Aliases

c_syntax_node_callback_t

@brief Function signature for a callback providing access to a syntax_node_t @ingroup tokenizer_and_parser_group @param[in] node The syntax_node_t being provided. This node should not be modified or freed by the callback. @param[in] context The context state pointer initially passed to the upstream function initiating the callback.