hyperonc::metta

Function get_atom_types

#[no_mangle]
pub extern "C" fn get_atom_types(
    space: *const space_t,
    atom: *const atom_ref_t,
    callback: c_atom_vec_callback_t,
    context: *mut c_void,
)

@brief Provides all types for atom in the context of space @ingroup metta_language_group @param[in] space A pointer to the space_t representing the space context in which to access the Atom’s types @param[in] atom A pointer to the atom_t or atom_ref_t representing the atom whose Types the function will access @param[in] callback A function that will be called to provide a vector of all type atoms associated with the atom argument atom @param[in] context A pointer to a caller-defined structure to facilitate communication with the callback function @note %Undefined% will be provided if atom has no type assigned. An empty vector will be provided if atom is a function call but expected types of arguments are not compatible with passed values