hyperonc::metta

Function metta_load_module_direct

#[no_mangle]
pub extern "C" fn metta_load_module_direct(
    metta_ref: *mut metta_t,
    name_ref: *const c_char,
    loader_mov: *mut module_loader_t,
) -> module_id_t

@brief Loads a module directly into the runner, from a module_loader_t @ingroup interpreter_group @param[in] metta_ref A pointer to the handle specifying the runner into which to load the module @param[in] name_ref A C-string specifying a name for the module @param[in] loader_mov The module_loader_t instance to load the module @return The module_id_t for the loaded module, or invalid if there was an error @note This function might be useful to provide MeTTa modules that are built-in as part of your application @note If this function encounters an error, the error may be accessed with metta_err_str()