hyperonc::metta

Function runner_state_new_with_atoms

#[no_mangle]
pub extern "C" fn runner_state_new_with_atoms(
    metta: *const metta_t,
    atoms: *const atom_vec_t,
) -> runner_state_t

@brief Creates a runner_state_t, to use for step-wise execution of a list of atoms @ingroup interpreter_group @param[in] metta A pointer to the runner handle in which to perform the run @param[in] atoms A pointer to an atom_vec_t containing the atoms to run @return The newly created runner_state_t, which can begin evaluating MeTTa code @warning The referenced atoms atom_vec_t must not be modified nor freed while the runner_state_t remains active @note The returned runner_state_t handle must be freed with runner_state_free()