hyperonc::metta

Function sexpr_parser_err_str

#[unsafe(no_mangle)]
pub extern "C" fn sexpr_parser_err_str(
    parser: *const sexpr_parser_t,
) -> *const c_char

@brief Returns the error string associated with the last sexpr_parser_parse call @ingroup tokenizer_and_parser_group @param[in] parser A pointer to the Parser, which is associated with the text to parse @return A pointer to the C-string containing the parse error that occurred, or NULL if no parse error occurred @warning The returned pointer should NOT be freed. It must never be accessed after the sexpr_parser_t has been freed, or any subsequent call to sexpr_parser_parse or sexpr_parser_parse_to_syntax_tree has been made.