hyperonc::metta

Function sexpr_parser_parse

#[unsafe(no_mangle)]
pub extern "C" fn sexpr_parser_parse(
    parser: *mut sexpr_parser_t,
    tokenizer: *const tokenizer_t,
) -> atom_t

@brief Parses the text associated with an sexpr_parser_t, and creates the corresponding Atom @ingroup tokenizer_and_parser_group @param[in] parser A pointer to the Parser, which is associated with the text to parse @param[in] tokenizer A pointer to the Tokenizer, to use to interpret atoms within the expression @return The new atom_t, which may be an Expression atom with many child atoms. Returns a none atom if parsing is finished, or an error expression atom if a parse error occurred. @note The caller must take ownership responsibility for the returned atom_t, and ultimately free it with atom_free() or pass it to another function that takes ownership responsibility @note If this function encounters an error, the error may be accessed with sexpr_parser_err_str()