hyperonc::metta

Function sexpr_parser_parse_to_syntax_tree

#[unsafe(no_mangle)]
pub extern "C" fn sexpr_parser_parse_to_syntax_tree(
    parser: *mut sexpr_parser_t,
) -> syntax_node_t

@brief Parses the text associated with an sexpr_parser_t, and creates a syntax tree @ingroup tokenizer_and_parser_group @param[in] parser A pointer to the Parser, which is associated with the text to parse @return The new syntax_node_t representing the root of the parsed tree @note The caller must take ownership responsibility for the returned syntax_node_t, and ultimately free it with syntax_node_free()