hyperonc::metta

Function validate_atom

#[unsafe(no_mangle)]
pub extern "C" fn validate_atom(
    space: *const space_t,
    atom: *const atom_ref_t,
) -> bool

@brief Checks whether atom is correctly typed @ingroup metta_language_group @param[in] space A pointer to the space_t representing the space context in which to perform the check @param[in] atom A pointer to the atom_t or atom_ref_t representing the atom whose Type the function will check @return true if the Atom is correctly typed, otherwise false @note This function can be used to check if function arguments have correct types