hyperonc::metta

Type Alias c_syntax_node_callback_t

pub type c_syntax_node_callback_t = extern "C" fn(node: *const syntax_node_t, context: *mut c_void);

@brief Function signature for a callback providing access to a syntax_node_t @ingroup tokenizer_and_parser_group @param[in] node The syntax_node_t being provided. This node should not be modified or freed by the callback. @param[in] context The context state pointer initially passed to the upstream function initiating the callback.