hyperonc

Module space

Structs

RustSpaceCommonData

space_api_t

@struct space_api_t @brief A table of functions to define the behavior of a Space implemented in C @ingroup space_impl_group @see space_new

space_event_t

@struct space_event_t @brief Represents a Space Event @ingroup space_observer_group @note Space Events can be read by observers monitoring a space, or created by custom by custom Space implementations, to communicate activity within the Space to observers

space_observer_api_t

@struct space_observer_api_t @brief A table of callback functions to define the behavior of a SpaceObserver implemented in C @ingroup space_observer_group @see space_register_observer

space_observer_t

@struct space_observer_t @brief Represents a Space Observer, registered with a Space @ingroup space_observer_group

space_params_t

@struct space_params_t @brief Data associated with this particular space, including the space’s payload and observers @ingroup space_impl_group

space_t

@struct space_t @brief A Space handle, providing access to a Space in which atoms may exist in relation to other atoms @ingroup space_client_group @note Multiple space_t handles may refer to the same underlying space. An underlying space will only be deallocated when all space_t handles that refer to it have been freed @note space_t must be freed with space_free(), or passed by value to a function that takes ownership of the space.

Enums

space_event_field_t

@brief Accessor constants, to access the fields of a space_event_t @ingroup space_observer_group

space_event_type_t

@brief Represents different types of Space Events @ingroup space_observer_group

Functions

space_add

@brief Adds an atom to the Space @ingroup space_client_group @param[in] space A pointer to the space_t handle to access @param[in] atom An atom_t representing the Atom to add @warning This function takes ownership of the supplied atom, and it should not be freed or accessed after it has been provided to this function

space_atom_count

@brief Returns the number of top-level atoms in a Space, if it can be readily determined @ingroup space_client_group @param[in] space A pointer to the space_t handle to access @return The number of top-level atoms in the Space, or -1 for spaces in which it is impossible or impractical to determine the number of atoms

space_clone_handle

@brief Creates a new space_t handle that refers to the same underlying space as another space_t @ingroup space_client_group @param[in] space A pointer to the space_t handle to clone @return A newly created space_t handle, providing access to the same underlying Space @note Each space_t is an independent object and must be individually freed, but both space_t objects reference the same underlying space, which is only deallocated when all references to it have been freed. @note The caller must take ownership responsibility for the returned space_t, and free it with space_free()

space_eq

@brief Checks if two space_t handles refer to the same underlying Space @ingroup space_client_group @param[in] a A pointer to the first space_t @param[in] b A pointer to the second space_t @return true is the underlying Space is the same for both space_t objects, and false otherwise

space_event_free

@brief Frees a space_event_t @ingroup space_impl_group @param[in] event The space_event_t to free

space_event_get_field_atom

@brief Accesses the atom associated with a field of a space_event_t @ingroup space_observer_group @param[in] event A pointer to the event to access @param[in] field A space_event_field_t specifying which field to access @return An atom_ref_t referencing the specified atom within the event @warning The returned atom_ref_t is borrowed from the space_event_t, and it must not be modified or accessed after the event has been freed

space_event_get_type

@brief Gets the type of a Space Event @ingroup space_observer_group @param[in] event A pointer to the event to inspect @return The type of the event

space_event_new_add

@brief Creates a new space_event_t representing an Add event @ingroup space_impl_group @param[in] atom The atom that is being added to the Space, to embed into the event @return The newly created space_event_t @note The caller must take ownership responsibility for the returned space_event_t and it must be freed with space_event_free() @warning This function takes ownership of the atom parameter, so it must not be subsequently access or freed

space_event_new_remove

@brief Creates a new space_event_t representing a Remove event @ingroup space_impl_group @param[in] atom The atom that is being removed from the Space, to embed into the event @return The newly created space_event_t @note The caller must take ownership responsibility for the returned space_event_t and it must be freed with space_event_free() @warning This function takes ownership of the atom parameter, so it must not be subsequently access or freed

space_event_new_replace

@brief Creates a new space_event_t representing a Replace event @ingroup space_impl_group @param[in] pattern The atom that is being matched in the Space, to embed into the event @param[in] tmpl The atom that is being used to construct new atoms in the Space, to embed into the event @return The newly created space_event_t @note The caller must take ownership responsibility for the returned space_event_t and it must be freed with space_event_free() @warning This function takes ownership of both the pattern and the tmpl parameter, so neither may be subsequently access nor freed

space_free

@brief Frees a space_t handle @ingroup space_client_group @param[in] space The handle to free @note The underlying Space may be deallocated if all handles that refer to it have been freed, otherwise the Space itself won’t be freed

space_get_payload

@brief Access the payload object belonging to a space implemented in C @ingroup space_client_group @param[in] space A pointer to the space_t handle to access @return The pointer to the payload object orginally used to create the Space @note This function is only valid for Spaces implemented via the HyperonC API @note This API is likely to change in the future to implement multi-threading as the space’s payload object must not be modified while other operations can access the space @warning The returned payload ptr must not be freed, nor may it be accessed after the space has been freed or modified

space_iterate

@brief Iterates all top-level Atoms in a Space, if that is possible @ingroup space_client_group @param[in] space A pointer to the space_t handle to access @param[in] callback A function that will be called for each top-level Atom in the Space @param[in] context A pointer to a caller-defined structure to facilitate communication with the callback function @return true if the space was sucessfully iterated, or false if the space does not support iteration

space_new

@brief Creates a new Space, backed by an implementation in C @ingroup space_client_group @param[in] api A pointer to the table of functions that implement the Space’s behavior @param[in] payload A pointer to a buffer to back the new Space @return A newly created space_t handle, to access the Space @warning This function takes ownership of the payload buffer, and it should not be freed or accessed after it has been provided to this function @note The caller must take ownership responsibility for the returned space_t, and free it with space_free()

space_new_grounding_space

@brief Creates a new Space, backed by a GroundSpace @ingroup space_client_group @return a space_t handle to the newly created Grounding Space @note The caller takes ownership responsibility for the returned space_t, and it must be freed with space_free()

space_observer_free

@brief Frees a space_observer_t @ingroup space_observer_group @param[in] observer The space_observer_t to free @note Freeing a Space Observer will atomatically de-register it with its associated space

space_observer_get_payload

@brief Returns a pointer to the payload associated with the space_observer_t @ingroup space_observer_group @param[in] observer A pointer to the space_observer_t in which to access the payload @return A pointer to the payload object associated with the Space Observer @warning The returned pointer must not be accessed after the space_observer_t has been freed, or after any operations have occurred that may have caused events to occur in the associated space @warning The returned pointer should never be freed directly. Call space_observer_free() when you are finished with the observer

space_params_notify_all_observers

@brief Notifies all associated observers of an event @ingroup space_impl_group @param[in] params A pointer to the space’s params @param[in] event A pointer to the event, to broadcast to all observers of the Space

space_query

@brief Queries a Space for atoms matching a pattern @ingroup space_client_group @param[in] space A pointer to the space_t handle to access @param[in] pattern A pointer to an atom_t or atom_ref_t to specify the pattern to match within the Space @return A bindings_set_t representing all possible results of the match @note The caller must take ownership responsibility for the returned bindings_set_t, and free it with bindings_set_free()

space_register_observer

@brief Registers a new observer, to monitor activity within the Space @ingroup space_observer_group @param[in] space A pointer to the space_t handle of the space to observe @param[in] observer_api A pointer to the table of functions that implement the observer’s behavior @param[in] observer_payload A pointer to a caller-defined object usable by the observer’s implementation functions @return A space_observer_t created to observe the space @note The caller must take ownership responsibility for the returned space_observer_t, and it must be freed with space_observer_free() @warning This function takes ownership of the observer_payload, and it should not be freed after it has been provided to this function

space_remove

@brief Removes a specific atom from the Space @ingroup space_client_group @param[in] space A pointer to the space_t handle to access @param[in] atom A pointer to an atom_t or atom_ref_t to specifying the atom to remove from the Space @return true if the atom was found and removed from the Space, false otherwise

space_replace

@brief Replaces an Atom in the Space with another Atom @ingroup space_client_group @param[in] space A pointer to the space_t handle to access @param[in] from A pointer to an atom_t or atom_ref_t to specify the atom to replace @param[in] to An atom_t to provide a new Atom, to replace the from atom in the Space @return true if an Atom was replaced in the Space, false otherwise @note The to atom will only be added to the Space if the from atom is found. Regardless, this functino will take ownership of the to atom. @warning This function takes ownership of the to atom, and it should not be freed or accessed after it has been provided to this function

space_subst

@brief Substitutes all Atoms matching a pattern with Atoms constructed from a template @ingroup space_client_group @param[in] space A pointer to the space_t handle to access @param[in] pattern A pointer to an atom_t or atom_ref_t to match Atom(s) in the Space @param[in] templ A pointer to an atom_t or atom_ref_t to specify a template from which to construct the new atoms @param[in] callback A function that will be called to provide access to a vec of newly created atoms @param[in] context A pointer to a caller-defined structure to facilitate communication with the callback function