hyperonc::space

Function space_clone_handle

#[unsafe(no_mangle)]
pub extern "C" fn space_clone_handle(space: *const space_t) -> space_t

@brief Creates a new space_t handle that refers to the same underlying space as another space_t @ingroup space_client_group @param[in] space A pointer to the space_t handle to clone @return A newly created space_t handle, providing access to the same underlying Space @note Each space_t is an independent object and must be individually freed, but both space_t objects reference the same underlying space, which is only deallocated when all references to it have been freed. @note The caller must take ownership responsibility for the returned space_t, and free it with space_free()