hyperonc::space

Function space_event_new_replace

#[no_mangle]
pub extern "C" fn space_event_new_replace(
    pattern: atom_t,
    tmpl: atom_t,
) -> space_event_t

@brief Creates a new space_event_t representing a Replace event @ingroup space_impl_group @param[in] pattern The atom that is being matched in the Space, to embed into the event @param[in] tmpl The atom that is being used to construct new atoms in the Space, to embed into the event @return The newly created space_event_t @note The caller must take ownership responsibility for the returned space_event_t and it must be freed with space_event_free() @warning This function takes ownership of both the pattern and the tmpl parameter, so neither may be subsequently access nor freed