hyperonc::space

Function space_get_payload

#[unsafe(no_mangle)]
pub extern "C" fn space_get_payload(
    space: *mut space_t,
) -> *mut c_void

@brief Access the payload object belonging to a space implemented in C @ingroup space_client_group @param[in] space A pointer to the space_t handle to access @return The pointer to the payload object orginally used to create the Space @note This function is only valid for Spaces implemented via the HyperonC API @note This API is likely to change in the future to implement multi-threading as the space’s payload object must not be modified while other operations can access the space @warning The returned payload ptr must not be freed, nor may it be accessed after the space has been freed or modified