hyperonc::space

Function space_iterate

#[unsafe(no_mangle)]
pub extern "C" fn space_iterate(
    space: *const space_t,
    callback: c_atom_callback_t,
    context: *mut c_void,
) -> bool

@brief Iterates all top-level Atoms in a Space, if that is possible @ingroup space_client_group @param[in] space A pointer to the space_t handle to access @param[in] callback A function that will be called for each top-level Atom in the Space @param[in] context A pointer to a caller-defined structure to facilitate communication with the callback function @return true if the space was sucessfully iterated, or false if the space does not support iteration