hyperonc::space

Function space_params_notify_all_observers

#[no_mangle]
pub extern "C" fn space_params_notify_all_observers(
    params: *const space_params_t,
    event: *const space_event_t,
)

@brief Notifies all associated observers of an event @ingroup space_impl_group @param[in] params A pointer to the space’s params @param[in] event A pointer to the event, to broadcast to all observers of the Space