hyperonc::space

Function space_subst

#[no_mangle]
pub extern "C" fn space_subst(
    space: *const space_t,
    pattern: *const atom_ref_t,
    templ: *const atom_ref_t,
    callback: c_atom_vec_callback_t,
    context: *mut c_void,
)

@brief Substitutes all Atoms matching a pattern with Atoms constructed from a template @ingroup space_client_group @param[in] space A pointer to the space_t handle to access @param[in] pattern A pointer to an atom_t or atom_ref_t to match Atom(s) in the Space @param[in] templ A pointer to an atom_t or atom_ref_t to specify a template from which to construct the new atoms @param[in] callback A function that will be called to provide access to a vec of newly created atoms @param[in] context A pointer to a caller-defined structure to facilitate communication with the callback function