hyperonc::metta

Function env_builder_use_default

#[no_mangle]
pub extern "C" fn env_builder_use_default() -> env_builder_t

@brief Creates an env_builder_t to specify that the default common environment should be used @ingroup environment_group @return The env_builder_t object specifying the common environment @note This function exists to supply an argument to metta_new_with_space when no special behavior is desired @note The env_builder_t must be passed to metta_new_with_space