hyperonc::metta

Function interpret_init

#[unsafe(no_mangle)]
pub extern "C" fn interpret_init(
    space: *mut space_t,
    expr: *const atom_ref_t,
) -> step_result_t

@brief Initializes an interpreter operation and take the initial step @ingroup interpreter_group @param[in] space A pointer to the Space in which to perform the operation @param[in] expr A pointer to an atom_t or atom_ref_t Expression atom to interpret @return A step_result_t representing the outcome from the initial step @note The returned value may represent an error, an immediate result value, or it may be necessary to call interpret_step() in a loop to fully evaluate the execution plan. Ultimately step_get_result() must be called to release the returned step_result_t