hyperonc::metta

Function metta_new_with_stdlib_loader

#[no_mangle]
pub extern "C" fn metta_new_with_stdlib_loader(
    stdlib_loader_mov: *mut module_loader_t,
    space_ref: *mut space_t,
    env_builder_mov: env_builder_t,
) -> metta_t

@brief Creates a new top-level MeTTa Runner, with the specified stdlib module loaded @ingroup interpreter_group @param[in] space_ref A pointer to a handle for the Space for use in the Runner’s top-level module @param[in] env_builder_mov An env_builder_t handle to configure the environment to use @param[in] stdlib_loader_mov Stdlib loader implemented in C code. Pass NULL to use the default stdlib @return A metta_t handle to the newly created Runner @note The caller must take ownership responsibility for the returned metta_t, and free it with metta_free() @note Most callers can simply call metta_new. This function is provided to support languages with their own stdlib, that needs to be loaded before the init.metta file is run