hyperonc

Module module

Structs

CModuleLoader

The wrapper of the module_loader_t providing Rust API of the C implementation

fs_module_format_t

@struct fs_module_format_t @brief A table of functions to load MeTTa modules from an arbitrary format @ingroup module_group @warning All of the functions in this interface may be called from threads outside the main thread, and may be called concurrently. Therefore these functions must be fully reentrant.

metta_mod_ref_t

@brief MettaMod C API wrapper @ingroup module_group

module_descriptor_t

@brief Identifies the properties on a specific module, including its name and version @ingroup module_group @note module_descriptor_t objects must be freed with module_descriptor_free or passed to a function that assumes responsibility for freeing them

module_id_t

@brief Identifies a loaded module inside a specific metta_t MeTTa runner @ingroup module_group @note It is not necessary to free module_id_t types

module_loader_t

@brief A C representation of the Rust [ModuleLoader] interface. User can provide [ModuleLoader] methods implemented in C. @ingroup module_group

Functions

metta_mod_ref_tokenizer

@brief Returns tokenizer of the MeTTa module @ingroup module_group @param[in] mod_ref A pointer to the metta_mod_ref_t @return tokenizer_t instance

module_descriptor_clone

@brief Creates a new module_descriptor_t that is a clone of the argument passed @ingroup module_group @param[in] descriptor The module_descriptor_t to clone @return The new module_descriptor_t @note The returned module_descriptor_t must be freed with module_descriptor_free()

module_descriptor_free

@brief Frees a module_descriptor_t @ingroup module_group @param[in] descriptor The module_descriptor_t to free

module_descriptor_new

@brief Creates a new module_descriptor_t with the specified name @ingroup module_group @param[in] name A C-style string containing the name of the module @return The new module_descriptor_t @note The returned module_descriptor_t must be freed with module_descriptor_free()

module_descriptor_new_with_path_and_fmt_id

module_id_is_valid

@brief Returns true is a module_id_t is valid, otherwise returns false @ingroup module_group @param[in] mod_id A pointer to the module_id_t to test for validity @return true if the module_id_t is valid, otherwise returns false