hyperonc::space

Function space_register_observer

#[no_mangle]
pub extern "C" fn space_register_observer(
    space: *mut space_t,
    observer_api: *const space_observer_api_t,
    observer_payload: *mut c_void,
) -> space_observer_t

@brief Registers a new observer, to monitor activity within the Space @ingroup space_observer_group @param[in] space A pointer to the space_t handle of the space to observe @param[in] observer_api A pointer to the table of functions that implement the observer’s behavior @param[in] observer_payload A pointer to a caller-defined object usable by the observer’s implementation functions @return A space_observer_t created to observe the space @note The caller must take ownership responsibility for the returned space_observer_t, and it must be freed with space_observer_free() @warning This function takes ownership of the observer_payload, and it should not be freed after it has been provided to this function