hyperonc

Module util

Structs

CWrite

Rust wrapper of reference of the std::fmt::Write implementer which is used to writer to the C code and get the result back. Typical use-case is implementing std::fmt::Display or std::fmt::Debug for the C object. See [crate::metta::module_loader_t].

write_t

@struct write_t @brief A handle to a Rust std::fmt::Writer to be used from C code @ingroup misc_group

Functions

cstr_as_str

cstr_into_string

log_error

@brief Logs an error through the MeTTa library’s logger. Does not halt execution. @ingroup misc_group @param[in] msg A C-style string containing the message to log

log_info

@brief Logs an informative message through the MeTTa library’s logger @ingroup misc_group @param[in] msg A C-style string containing the message to log

log_warn

@brief Logs a warning through the MeTTa library’s logger @ingroup misc_group @param[in] msg A C-style string containing the message to log

str_as_cstr

string_as_cstr

write_str

@brief Write C string into a Rust writer @ingroup misc_group @param[in] cwrite A handle to a Rust writer @param[in] text C string in UTF-8 format to be written @return 0 if string is written successfully, non-zero otherwise